Organize opinions on default spines
