+++ src/operators/consolidate.rs
@@ -48,8 +48,8 @@ where
     /// }
     /// ```
     pub fn consolidate(&self) -> Self {
-        use trace::implementations::OrdKeySpine as DefaultKeyTrace;
-        self.consolidate_named::<DefaultKeyTrace<_,_,_>>("Consolidate")
+        use trace::implementations::KeySpine;
+        self.consolidate_named::<KeySpine<_,_,_>>("Consolidate")
     }
 
     /// As `consolidate` but with the ability to name the operator and specify the trace type.
