+++ src/trace/implementations/mod.rs
@@ -45,8 +45,125 @@ pub(crate) mod merge_batcher_col;
 
 pub use self::merge_batcher::MergeBatcher as Batcher;
 
-mod ord;
+pub mod ord;
 pub mod ord_neu;
 
-pub use self::ord::{OrdKeySpine, ColKeySpine};
-pub use self::ord_neu::{OrdValSpine, ColValSpine};
+// Opinionated takes on default spines.
+pub use self::ord::OrdValSpine as ValSpine;
+pub use self::ord::OrdKeySpine as KeySpine;
+
+use timely::container::columnation::{Columnation, TimelyStack};
+use lattice::Lattice;
+use difference::Semigroup;
+use trace::layers::BatchContainer;
+use trace::layers::ordered::OrdOffset;
+
+/// A type that names constituent update types.
+pub trait Update {
+    /// Key by which data are grouped.
+    type Key: Ord+Clone;
+    /// Values associated with the key.
+    type Val: Ord+Clone;
+    /// Time at which updates occur.
+    type Time: Ord+Lattice+timely::progress::Timestamp+Clone;
+    /// Way in which updates occur.
+    type Diff: Semigroup+Clone;
+}
+
+impl<K,V,T,R> Update for ((K, V), T, R)
+where
+    K: Ord+Clone,
+    V: Ord+Clone,
+    T: Ord+Lattice+timely::progress::Timestamp+Clone,
+    R: Semigroup+Clone,
+{
+    type Key = K;
+    type Val = V;
+    type Time = T;
+    type Diff = R;
+}
+
+/// A type with opinions on how updates should be laid out.
+pub trait Layout {
+    /// The represented update.
+    type Target: Update;
+    /// Offsets to use from keys into vals.
+    type KeyOffset: OrdOffset;
+    /// Offsets to use from vals into updates.
+    type ValOffset: OrdOffset;
+    /// Container for update keys.
+    type KeyContainer:
+        RetainFrom<<Self::Target as Update>::Key>+
+        BatchContainer<Item=<Self::Target as Update>::Key>;
+    /// Container for update vals.
+    type ValContainer:
+        RetainFrom<<Self::Target as Update>::Val>+
+        BatchContainer<Item=<Self::Target as Update>::Val>;
+    /// Container for update vals.
+    type UpdContainer:
+        BatchContainer<Item=(<Self::Target as Update>::Time, <Self::Target as Update>::Diff)>;
+}
+
+/// A layout that uses vectors
+pub struct Vector<U: Update, O: OrdOffset = usize> {
+    phantom: std::marker::PhantomData<(U, O)>,
+}
+
+impl<U: Update+Clone, O: OrdOffset> Layout for Vector<U, O> {
+    type Target = U;
+    type KeyOffset = O;
+    type ValOffset = O;
+    type KeyContainer = Vec<U::Key>;
+    type ValContainer = Vec<U::Val>;
+    type UpdContainer = Vec<(U::Time, U::Diff)>;
+}
+
+/// A layout based on timely stacks
+pub struct TStack<U: Update, O: OrdOffset = usize> {
+    phantom: std::marker::PhantomData<(U, O)>,
+}
+
+impl<U: Update+Clone, O: OrdOffset> Layout for TStack<U, O>
+where
+    U::Key: Columnation,
+    U::Val: Columnation,
+    U::Time: Columnation,
+    U::Diff: Columnation,
+{
+    type Target = U;
+    type KeyOffset = O;
+    type ValOffset = O;
+    type KeyContainer = TimelyStack<U::Key>;
+    type ValContainer = TimelyStack<U::Val>;
+    type UpdContainer = TimelyStack<(U::Time, U::Diff)>;
+}
+
+/// A container that can retain/discard from some offset onward.
+pub trait RetainFrom<T> {
+    /// Retains elements from an index onwards that satisfy a predicate.
+    fn retain_from<P: FnMut(usize, &T)->bool>(&mut self, index: usize, predicate: P);
+}
+
+impl<T> RetainFrom<T> for Vec<T> {
+    fn retain_from<P: FnMut(usize, &T)->bool>(&mut self, index: usize, mut predicate: P) {
+        let mut write_position = index;
+        for position in index .. self.len() {
+            if predicate(position, &self[position]) {
+                self.swap(position, write_position);
+                write_position += 1;
+            }
+        }
+        self.truncate(write_position);
+    }
+}
+
+impl<T: Columnation> RetainFrom<T> for TimelyStack<T> {
+    fn retain_from<P: FnMut(usize, &T)->bool>(&mut self, index: usize, mut predicate: P) {
+        let mut position = index;
+        self.retain_from(index, |item| {
+            let result = predicate(position, item);
+            position += 1;
+            result
+        })
+    }
+}
